scripts/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
azure-cognitiveservices-speech>=1.34.0
